:ticket: metadata added to all pages
