+++ app/home/room/[roomId]/create-prize/page.tsx
@@ -1,106 +1,19 @@
-"use client";
-import { createPrize } from "@/api/prize";
-import { getRoom } from "@/api/room";
-import PrimaryButton from "@/components/common/button/PrimaryButton";
-import ComponentTitle from "@/components/common/ComponentTitle";
-import ComponentWithHeader from "@/components/common/ComponentWithHeader";
-import Loader from "@/components/common/Loader";
-import CustomInputField from "@/components/form/CustomInputField";
-import RoomName from "@/components/room/RoomName";
-import { IRoomBasic } from "@/types/IRoom.model";
-import { Form, Formik } from "formik";
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { toast } from "react-toastify";
+import RoomCreatePrizeContent from "@/components/room/pages/RoomCreatePrizeContent";
+import { Metadata } from "next";
 
-interface ICreatePrizeProps {
+export const metadata: Metadata = {
+  title: "Guessers | Room | Create Prize",
+  description: "Create prizes to be bought",
+};
+
+interface IRoomCreatePrizeProps {
   params: {
     roomId: string;
   };
 }
 
-const CreatePrize = ({ params }: ICreatePrizeProps) => {
-  const [room, setRoom] = useState<IRoomBasic>();
-  const [loading, setLoading] = useState(false);
-  const router = useRouter();
-
-  const initialValues = {
-    name: "",
-    description: "",
-    value: 0,
-  };
-
-  useEffect(() => {
-    getRoom(params.roomId).then((response) => {
-      setRoom(response.data);
-    });
-  }, [params.roomId]);
-
-  return (
-    room && (
-      <div className="flex flex-col justify-center items-center">
-        <ComponentTitle text={"Create Prize"} />
-
-        <RoomName roomName={room.name} roomId={room.id} />
-
-        <Formik
-          initialValues={initialValues}
-          onSubmit={(values) => {
-            if (values.name === "" || values.description === "") {
-              toast.error("Please fill all fields");
-              return;
-            }
-            setLoading(true);
-            createPrize({
-              createPrizeRequest: values,
-              roomId: params.roomId,
-            })
-              .then(() => {
-                setTimeout(() => {
-                  router.push(`/home/room/${params.roomId}/guess`);
-                }, 1000);
-              })
-              .finally(() => {
-                setLoading(false);
-              });
-          }}
-        >
-          <Form className="flex flex-col justify-center items-center">
-            <CustomInputField
-              name={`name`}
-              type="text"
-              placeholder={"name"}
-              withLabel={true}
-            />
-
-            <CustomInputField
-              name={`description`}
-              type="text"
-              placeholder={"description"}
-              withLabel={true}
-            />
-
-
-            <CustomInputField
-              name={`value`}
-              type="select"
-              placeholder={"value"}
-              withLabel={true}
-              options={[{ value: 100, label: "100" }, { value: 200, label: "200" }, { value: 500, label: "500" }, { value: 1000, label: "1000" }, { value: 2000, label: "2000" }, { value: 5000, label: "5000" }, { value: 10000, label: "10000" }]}
-            />
-            
-            <div className="my-2">
-              {loading ? (
-                <Loader />
-              ) : (
-                <PrimaryButton type="submit" text={"Create Prize"} bg />
-              )}
-            </div>
-          </Form>
-        </Formik>
-      </div>
-    )
-  );
+const RoomCreatePrize = ({ params }: IRoomCreatePrizeProps) => {
+  return <RoomCreatePrizeContent params={params} />;
 };
 
-export default CreatePrize;
+export default RoomCreatePrize;
