+++ app/home/room/[roomId]/event/[eventId]/finalize/page.tsx
@@ -1,13 +1,10 @@
-"use client";
-import { finalizeEvent, getEvent } from "@/api/event";
-import PrimaryButton from "@/components/common/button/PrimaryButton";
-import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
-import { IEvent, IEventGuessOptionCase, IFinalizeEventRequest } from "@/types/IEvent.model";
-import ComponentTitle from "@/components/common/ComponentTitle";
-import { IRoomBasic } from "@/types/IRoom.model";
-import { getRoom } from "@/api/room";
-import RoomName from "@/components/room/RoomName";
+import EventFinalizeContent from "@/components/event/pages/EventFinalizeContent";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Guessers | Room | Event Finalize",
+  description: "Finalize an event for your room",
+};
 
 interface IEventFinalizeProps {
   params: {
@@ -17,118 +14,7 @@ interface IEventFinalizeProps {
 }
 
 const EventFinalize = ({ params }: IEventFinalizeProps) => {
-  const [room, setRoom] = useState<IRoomBasic>();
-  const [event, setEvent] = useState<IEvent>();
-  const [winningOptions, setWinningOptions] = useState<IEventGuessOptionCase[]>(
-    []
-  );
-  const router = useRouter();
-
-  useEffect(() => {
-    getEvent(params.eventId).then((response) => {
-      setEvent(response.data);
-    });
-  }, []);
-
-  useEffect(() => {
-    getRoom(params.roomId).then((response) => {
-      setRoom(response.data);
-    });
-  }, [params.roomId]);
-
-  const handleChangeWinningOptions = (
-    eventGuessOptionCase: IEventGuessOptionCase
-  ) => {
-    if (winningOptions.includes(eventGuessOptionCase)) {
-      setWinningOptions(
-        winningOptions.filter(
-          (winningOption) => winningOption !== eventGuessOptionCase
-        )
-      );
-    } else {
-      setWinningOptions([...winningOptions, eventGuessOptionCase]);
-    }
-  };
-
-  const handleFinalize = () => {
-    const winningCaseIds = winningOptions.map(
-      (winningOption) => winningOption.id
-    );
-    const finalizeRequest: IFinalizeEventRequest = {
-      winnerEventGuessOptionCases: winningCaseIds,
-    };
-    finalizeEvent({request: finalizeRequest, eventId: params.eventId, roomId: params.roomId}).finally(
-      () => {
-        router.push("/home/room/" + params.roomId + "/guess");
-      }
-    );
-  };
-  return (
-    event && room && (
-      <div className="flex flex-col justify-center items-center text-text-default text-xs">
-        <ComponentTitle text="Finalize Event" />
-
-        <RoomName roomName={room.name} roomId={room.id} />
-
-        <p>Select the winning options.</p>
-        <div className="flex flex-col justify-center items-center w-full rounder-md text-xs bg-background-bright border-2 border-primary-default rounded-md">
-          <div className="flex justify-between items-center w-full font-bold p-4 rounded-md">
-            <p>{event.name}</p>
-
-            <p>{new Date(event.createdOn).toLocaleString()}</p>
-          </div>
-
-          <div className="w-full grid grid-cols-2 gap-5 p-2">
-            {event.eventGuessOptions?.map((eventGuessOption) => (
-              <div
-                key={eventGuessOption.id}
-                className="flex flex-col justify-start items-start"
-              >
-                <p className="text-background-bright bg-primary-default p-2 rounded-md text-xs lowercase my-1 font-bold">
-                  {eventGuessOption.name}
-                </p>
-                <div className="w-full grid auto-cols-fr grid-flow-col">
-                  {eventGuessOption.eventGuessOptionCases?.map(
-                    (eventGuessOptionCase) => {
-                      return (
-                        <button
-                          key={
-                            eventGuessOptionCase.name +
-                            "_" +
-                            eventGuessOptionCase.odds
-                          }
-                          className="flex flex-col justify-center items-center hover:cursor-pointer m-0.5"
-                          onClick={() =>
-                            handleChangeWinningOptions(eventGuessOptionCase)
-                          }
-                        >
-                          <div className="text-text w-full flex justify-start items-center py-2">
-                            <p className="truncate font-bold">{eventGuessOptionCase.name}</p>
-                          </div>
-                          <div
-                            className={`${
-                              winningOptions.includes(eventGuessOptionCase)
-                                ? "bg-primary-default text-background-bright"
-                                : "text-primary-default bg-background-bright"
-                            } w-full py-2 m-1 flex justify-evenly items-center font-bold border-2 border-primary-default rounded-md`}
-                          >
-                            <p>{eventGuessOptionCase.odds}</p>
-                          </div>
-                        </button>
-                      );
-                    }
-                  )}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-        <div className="my-4">
-        <PrimaryButton type="submit" text={"Finalize Event"} onClick={() => handleFinalize()} bg />
-        </div>
-      </div>
-    )
-  );
+  return <EventFinalizeContent params={params} />;
 };
 
 export default EventFinalize;
