+++ app/home/room/[roomId]/invite/page.tsx
@@ -1,74 +1,17 @@
-"use client";
-import { getRoom, invitePeople } from "@/api/room";
-import PrimaryButton from "@/components/common/button/PrimaryButton";
-import ComponentTitle from "@/components/common/ComponentTitle";
-import Loader from "@/components/common/Loader";
-import CustomInputField from "@/components/form/CustomInputField";
-import RoomName from "@/components/room/RoomName";
-import { IRoomBasic } from "@/types/IRoom.model";
-import { Form, Formik } from "formik";
-import { useEffect, useState } from "react";
-import { toast } from "react-toastify";
+import RoomInviteContent from "@/components/room/pages/RoomInviteContent";
+import { Metadata } from "next";
 
-interface IInvitePeopleProps {
+export const metadata: Metadata = {
+  title: "Guessers | Room | Invite",
+  description: "Invite your friends to your room",
+};
+
+interface IRoomInviteProps {
   params: { roomId: string };
 }
 
-const InvitePeople = ({ params }: IInvitePeopleProps) => {
-  const [room, setRoom] = useState<IRoomBasic>();
-  const [loading, setLoading] = useState(false);
-  const initialValues = {
-    username: "",
-  };
-
-  useEffect(() => {
-    getRoom(params.roomId).then((response) => {
-      setRoom(response.data);
-    });
-  }, [params.roomId]);
-  return (
-    <div className="flex flex-col justify-center items-center my-6">
-      <ComponentTitle text={"Invite People"} />
-      {room?.name && <RoomName roomName={room.name} roomId={params.roomId} />}
-
-      <Formik
-        initialValues={initialValues}
-        onSubmit={(values) => {
-          if (values.username === null || values.username.length < 3) {
-            toast.error(
-              "Username is required and must be at least 3 characters long."
-            );
-            return;
-          }
-          setLoading(true);
-
-          invitePeople({
-            invitedUsername: values.username,
-            roomId: params.roomId,
-          }).finally(() => {
-            setLoading(false);
-          });
-        }}
-      >
-        <Form className="flex flex-col justify-center items-center">
-          <CustomInputField
-            name={`username`}
-            type="text"
-            placeholder={"username"}
-            withLabel={true}
-          />
-
-          <div className="my-2">
-          {loading ? (
-            <Loader />
-          ) : (
-            <PrimaryButton type="submit" text={"Invite"} bg />
-          )}
-          </div>
-        </Form>
-      </Formik>
-    </div>
-  );
+const RoomInvite = ({ params }: IRoomInviteProps) => {
+  return <RoomInviteContent params={params} />;
 };
 
-export default InvitePeople;
+export default RoomInvite;
