+++ app/home/room/[roomId]/layout.tsx
@@ -1,9 +1,11 @@
-"use client";
-import React, { useEffect, useState } from "react";
-import { getRoomUser } from "@/api/room";
-import RoomHeader from "@/components/room/RoomHeader";
-import Link from "next/link";
-import { usePathname } from "next/navigation";
+import RoomLayoutContent from "@/components/room/RoomLayoutContent";
+import { Metadata } from "next";
+import React from "react";
+
+export const metadata: Metadata = {
+  title: "Guessers | Room",
+  description: "Guessers.io, where you meet up and guess with your friends!",
+};
 
 interface IRoomLayoutProps {
   params: {
@@ -13,73 +15,7 @@ interface IRoomLayoutProps {
 }
 
 const RoomLayout = ({ params, children }: IRoomLayoutProps) => {
-  const [roomUser, setRoomUser] = useState(null);
-  const pathname = usePathname();
-
-  useEffect(() => {
-    const fetchRoomUser = async () => {
-      const userResponse = await getRoomUser(params.roomId);
-      setRoomUser(userResponse.data);
-    };
-
-    fetchRoomUser();
-  }, [params.roomId]);
-
-  if (!roomUser) return;
-  return (
-    <div>
-      {!pathname.endsWith("create") && !pathname.endsWith("invite") && !pathname.endsWith("create-prize") && !pathname.endsWith("lendtoken") && !pathname.endsWith("finalize") && ( // Do not show tabs in create room page
-        <>
-          <RoomHeader roomUser={roomUser} />
-
-          <div className="flex w-full text-sm gap-1">
-            <Link
-              className={`border-2 p-2 text-xs flex-1 flex justify-center items-center hover:border-b-primary transition-all font-bold ${
-                pathname.endsWith("guess")
-                  ? "border-primary-default bg-primary-default text-background-bright hover:bg-primary-bright"
-                  : "border-primary-default text-text-default bg-background-bright hover:bg-primary-default hover:text-background-bright"
-              }`}
-              href={`/home/room/${params.roomId}/guess`}
-            >
-              {"Guess"}
-            </Link>
-            <Link
-              className={`border-2 p-2 text-xs flex-1 flex justify-center items-center hover:border-b-primary transition-all font-bold ${
-                pathname.endsWith("ranks")
-                  ? "border-primary-default bg-primary-default text-background-bright hover:bg-primary-bright"
-                  : "border-primary-default text-text-default bg-background-bright hover:bg-primary-default hover:text-background-bright"
-              }`}
-              href={`/home/room/${params.roomId}/ranks`}
-            >
-              {"Ranks"}
-            </Link>
-            <Link
-              className={`border-2 p-2 text-xs flex-1 flex justify-center items-center hover:border-b-primary transition-all font-bold ${
-                pathname.endsWith("papers")
-                  ? "border-primary-default bg-primary-default text-background-bright hover:bg-primary-bright"
-                  : "border-primary-default text-text-default bg-background-bright hover:bg-primary-default hover:text-background-bright"
-              }`}
-              href={`/home/room/${params.roomId}/papers`}
-            >
-              {"Guess Papers"}
-            </Link>
-            <Link
-              className={`border-2 p-2 text-xs flex-1 flex justify-center items-center hover:border-b-primary transition-all font-bold ${
-                pathname.endsWith("prizes")
-                  ? "border-primary-default bg-primary-default text-background-bright hover:bg-primary-bright"
-                  : "border-primary-default text-text-default bg-background-bright hover:bg-primary-default hover:text-background-bright"
-              }`}
-              href={`/home/room/${params.roomId}/prizes`}
-            >
-              {"Prizes"}
-            </Link>
-          </div>
-        </>
-      )}
-
-      {children}
-    </div>
-  );
+  return <RoomLayoutContent params={params}>{children}</RoomLayoutContent>;
 };
 
 export default RoomLayout;
