+++ app/home/room/[roomId]/prizes/page.tsx
@@ -1,8 +1,10 @@
-"use client";
-import { getRoomPrizes } from "@/api/prize";
-import RoomActivePrizes from "@/components/room/RoomActivePrizes";
-import { IPaging } from "@/types/IRequest.model";
-import { useEffect, useState } from "react";
+import RoomPrizesContent from "@/components/room/pages/RoomPrizesContent";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Guessers | Room | Prizes",
+  description: "Buy a prize with your guessers token",
+};
 
 interface IRoomPrizesProps {
   params: {
@@ -11,43 +13,7 @@ interface IRoomPrizesProps {
 }
 
 const RoomPrizes = ({ params }: IRoomPrizesProps) => {
-  const [roomActivePrizes, setRoomActivePrizes] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [paging, setPaging] = useState<IPaging>({
-    page: 0,
-    size: 5,
-    totalPages: 0,
-    totalElements: 0,
-  });
-
-  const fetchRoomActivePrizes = async (roomId: string) => {
-    if (!roomId) return;
-
-    setLoading(true);
-    await getRoomPrizes({
-      roomId,
-      paging,
-      active: true,
-    })
-      .then((response) => {
-        setRoomActivePrizes(response.data.content);
-        setPaging({
-          page: response.data.page.number,
-          size: response.data.page.size,
-          totalPages: response.data.page.totalPages,
-          totalElements: response.data.page.totalElements,
-        });
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
-
-  useEffect(() => {
-    fetchRoomActivePrizes(params.roomId);
-  }, [params?.roomId, paging.page]);
-
-  return <RoomActivePrizes prizes={roomActivePrizes} paging={paging} setPaging={setPaging} loading={loading} />;
+  return <RoomPrizesContent params={params} />;
 };
 
 export default RoomPrizes;
