+++ app/page.tsx
@@ -37,10 +37,14 @@ function Home() {
     return (
       <div className="flex flex-col items-center justify-start text-text lg:text-sm text-xs w-full">
         <div className="my-4"></div>
-        <div className="flex justify-center items-center w-full">
-          <Image src="/logo/logo-with-text.svg" alt="logo" width={240} height={240} />
+        <div className="flex justify-center items-center w-full mb-4">
+          <Image
+            src="/logo/logo-with-text.svg"
+            alt="logo"
+            width={240}
+            height={240}
+          />
         </div>
-        <div className="my-4"></div>
         <Welcomer stats={stats} />
 
         <LoginOrRegister />
