+++ components/event/EventCard.tsx
